requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.7
action-tutorials-py==0.20.7
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
builtin-interfaces==1.2.2
certifi==2026.1.4
composition-interfaces==1.2.2
cv-basics==0.0.0
cv-bridge==3.2.1
demo-nodes-py==0.20.7
diagnostic-msgs==4.9.0
distro==1.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
exceptiongroup==1.3.1
gazebo-msgs==3.9.0
geometry-msgs==4.9.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
image-geometry==3.2.1
imap-tools==1.11.1
interactive-markers==2.3.2
laser-geometry==2.4.1
launch==1.0.13
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
lifecycle-msgs==1.2.2
logging-demo==0.20.7
map-msgs==2.1.0
message-filters==4.3.11
nav-msgs==4.9.0
notion-client==2.7.0
osrf-pycommon==2.1.6
pcl-msgs==1.0.0
pendulum-msgs==0.20.7
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-qt-binding==1.1.3
python-telegram-bot==22.5
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.7
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.7
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
sniffio==1.3.1
soupsieve==2.8.1
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.25.17
tf2-kdl==0.25.17
tf2-msgs==0.25.17
tf2-py==0.25.17
tf2-ros-py==0.25.17
tf2-sensor-msgs==0.25.17
tf2-tools==0.25.17
topic-monitor==0.20.7
trajectory-msgs==4.9.0
turtle-circle==0.0.0
turtlebot3-teleop==2.3.3
turtlesim==1.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
visualization-msgs==4.9.0
